Add command to open post at point, bind to RET by default
Opening posts from the *org-blog* buffer is probably the most complex part of
the system.

org-blog-mode is caching two data structures: the vector of posts pulled from
the server which is used to display a list of posts in the *org-blog*
buffer (stored as buffer local variable `posts-cache`) and a hash table which
maps from a filename to a full file path.

The hash table is also stored on the local file system at the path specified by
the `org-blog-local-data-file` configuration variable.

When a user posts a buffer to an org blog server, an entry is added to the hash
table so that the post's path can quickly be looked up when the user
wants to open the post from the *org-blog* buffer. The path of the post relative to the
user's home directory is also sent to the server to be stored with the post.

The path that's stored on the server is meant to make it easy to sync posts
across different machines. Consider the following use case:
1. I write a blog post in ~/blog/new_post.org and run `org-blog-post-buffer`
2. On a different computer, I run `org-blog` pointing to the same server
3. I see the new post in the list and attempt to open it for editing
4. Here there are two cases:
A. I'm using some type of file synchronization system (like SyncThing), so a
file at ~/blog/new_post.org already exists on the other computer. In this case,
org-blog-mode will find the file using the path from the server.
B. I don't have the post on the other computer, so I need to pull it down from
the server. In this case, org-blog-mode will prompt me as to where to store the
post (defaulting to the path from the server), get the raw post from the server,
and write it to the file at the specified path.

This is probably a sort of rare use-case, but it will make using org-blog-mode a
little bit more convenient, at least for me.

This use-case also got me thinking about trying to make it possible to use the
org-blog server as a git remote. This would be an alternative to using
org-blog-mode and would be a more generic solution. With this approach,
org-blog-mode is probably a bit redundant. magit and dired would be enough of an
interface to manage posts.